Opaleye.SOT

Description

For a better experience, it is recommended that you import this module unqualified as follows:

import Opaleye.SOT

Note that Opaleye.SOT re-exports all of Opaleye.SOT.Run, you might want to refer to that module for documentation.

Both Opaleye.SOT and Opaleye.SOT.Run override some of the names exported by the Opaleye, so it is recommended that you import Opaleye, if needed, qualified as:

import qualified Opaleye as O

This module doesn't export any infix operator.

Executing queries

module Opaleye.SOT.Run

Defining a Tabla

class ITabla t => Tabla t

Tabla means table in spanish.

An instance of this class can uniquely describe a PostgreSQL table and how to convert back and forth between it and its Haskell representation used when writing Opaleye queries.

The t type is only used as a tag for the purposes of uniquely identifying this Tabla.

Associated Types

type Database t :: *

Some kind of unique identifier used for telling appart the database where this table exists from other databases, so as to avoid accidentally mixing tables from different databases in queries.

type SchemaName t :: Symbol

PostgreSQL schema name where to find the table (defaults to "public", PostgreSQL's default schema name).

type TableName t :: Symbol

Table name.

type Cols t :: [Col Symbol WD RN * *]

Columns in this table. See the documentation for Col.

Working with Tabla

table :: Tabla t => T t -> Table (PgW t) (PgR t)

Build the Opaleye Table for a Tabla.

queryTabla :: Tabla t => T t -> Query (PgR t)

Like opaleye's queryTable, but for a Tabla.

type HsR t = Rec t (Cols_HsR t)

Expected output type for runQuery on a (PgR t).

Important: If you are expecting a (PgR t) on the right side of a leftJoin, you will need to use (Maybe (PgR t)).

Mnemonic: Haskell Read.

type HsI t = Rec t (Cols_HsI t)

HsI t is the Haskell representation of Haskell values to be inserted to the database, as taken by "Opaleye.SOT.Run.runInsertTabla".

An HsI t can always be converted to a PgW t using pgWfromHsI, in case you need that for with the more general "Opaleye.SOT.Run.runInsert".

Mnemonic: Haskell Insert.

mkHsI

Arguments

:: (Tabla t, HRearrange (LabelsOf (Cols_HsI t)) xs (Cols_HsI t))

=> T t

-> ((forall c a. C c -> a -> TCa t c a) -> HList xs)

-> HsI t

Helper function to safely build an HsI t.

The type of this function isn't easy to understand, but an example should clarify its usage. Here we asume we have a Person datatype and a TPerson which is an instance of Tisch.

personToHsI :: Person -> HsI TPerson
personToHsI (Person name age) =
   mkHsI (T :: TPerson) $ \set_ -> hBuild
       (set_ (C :: C "name") name)
       (set_ (C :: C "age") age)

You are not required to use this function to build an HsI t if working with the tools from Data.HList is sufficient for you. This is just a convenience.

type PgR t = Rec t (Cols_PgR t)

Output type of queryTabla (T t).

Mnemonic: PostGresql Read.

type PgRN t = Rec t (Cols_PgRN t)

Like (PgRN t) but every field is Koln, as in the output type of the right hand side of a leftJoin with '(table t).

Mnemonic: PostGresql Read Nulls.

type PgW t = Rec t (Cols_PgW t)

Representation of PostgreSQL values to be written to the database. This type can be used as input for "Opaleye.SOT.Run.runInsert" and similar.

An HsI t can always be converted to a PgW t using pgWfromHsI, in

Mnemonic: PostGresql Write.

pgWfromHsI :: Tabla t => HsI t -> PgW t

Convert a custom Haskell type to a representation appropiate for inserting it as a new row using runInsert.

pgWfromPgR :: Tabla t => PgR t -> PgW t

Convert a (PgR t) resulting from a queryTable-like operation to a (PgW t) that can be used in a runUpdate-like operation.

Kol

data Kol a

Like opaleye's (Column x), but with x guaranteed to be not Nullable. If you need to have a Nullable column type, use Koln instead.

Build using kol or Kol.

Notice that Kol is very different from Col: Col is used to describe the properties of a column at compile time. Kol is used at runtime for manipulating with values stored in columns.

We do not use (Column x), instead we use (Kol y) where x ~ PgType y. This is where we drift a bit appart from Opaleye. See https://github.com/tomjaguarpaw/haskell-opaleye/issues/97

Constructors

PgTyped a => Kol

Fields

unKol :: Column (PgType a)

kolCoerce :: KolCoerce a b => Kol a -> Kol b

unsafeCoerceKol :: (PgTyped a, PgTyped b) => Kol a -> Kol b

class PgPrimType p => ToKol a p where

Build a Kol.

You need to provide a ToKol instance for every Haskell type you plan to convert to its PostgreSQL representation as Kol.

A a default implementation of kol is available for Wrapped instances:

default kol :: (PgTyped b, PgType b ~ p, Wrapped a, ToKol (Unwrapped a) p) => a -> Kol b
kol = kol . view _Wrapped'

Implementation notice: This class overlaps in purpose with Opaleye's Constant. Technicaly, we don't really need to repeat those instances here: we could just rely on Opaleye's Constant. However, as of today, Opaleye's Constant provides some undesired support which we deliberately want to avoid here. Namely, we don't want to support converting Int to PGInt4. If this is fixed upstream, we might go back to relying on Constant if suitable. See https://github.com/tomjaguarpaw/haskell-opaleye/pull/110

Minimal complete definition

Nothing

Methods

kol :: (PgTyped b, PgType b ~ p) => a -> Kol (b :: kb)

Convert a constant Haskell value (say, a Bool) to its equivalent PostgreSQL representation as a (Kol PGBool).

Some example simplified types:

kol :: Bool -> Kol PGBool
kol :: Int32 -> Kol PGInt4

liftKol1

Arguments

:: (PgTyped a, PgTyped b, Op1 kol kol')

=> (Column (PgType a) -> Column (PgType b))

-> kol a -> kol' b

Converts an unary function on Opaleye's Column to an unary function taking any of Kol and Koln as argument, with the result type fully determined by it.

liftKol2 :: (PgTyped a, PgTyped b, PgTyped c, Op2 kol kol' kol'') => (Column (PgType a) -> Column (PgType b) -> Column (PgType c)) -> kol a -> kol' b -> kol'' c

Converts a binary function on Opaleye's Columns to an binary function taking any of Kol and Koln as arguments, with the result type fully determined by them.

Koln

data Koln a

Like opaleye's Column (Nullable x), but with x guaranteed to be not-Nullable.

Think of Koln a as Maybe (Kol a), with nul being analogous to the Nothing constructor and koln being analogous to the Just constructor.

Build safely using nul, koln, fromKol or Koln.

Notice that Koln is very different from Col: Col is used to describe the properties of a column at compile time. Koln is used at runtime for manipulating with values stored in columns.

We do not use Column (Nullable x), instead we use Koln y where x ~ PgType y. This is where we drift a bit appart from Opaleye. See https://github.com/tomjaguarpaw/haskell-opaleye/issues/97

Constructors

PgTyped a => Koln

Fields

unKoln :: Column (Nullable (PgType a))

koln :: (ToKol a p, PgTyped b, PgType b ~ p) => a -> Koln b

Build a Koln from a Haskell term. This is like the Just constructor for Maybe

nul :: PgTyped a => Koln a

PostgreSQL's NULL value. This is like the Nothing constructor for Maybe

class KolLike f => FromKol f where

Methods

fromKol' :: PgTyped a => Kol a -> f a

Like fromKol, but generalizing the return type to either Kol or Koln.

fromKol :: PgTyped a => Kol a -> Koln a

Convert a Kol to a Koln.

matchKoln :: (PgTyped a, PgTyped b) => Kol b -> (Kol a -> Kol b) -> Koln a -> Kol b

Case analysis for Koln. Like maybe for Maybe.

If (Koln a) is NULL, then evaluate to the first argument, otherwise it applies the given function to the underlying (Kol a).

isNull :: PgTyped a => Koln a -> Kol PGBool

Whether a Koln is nul (NULL).

mapKoln :: (PgTyped a, PgTyped b) => (Kol a -> Kol b) -> Koln a -> Koln b

Like fmap for Maybe.

Apply the given function to the underlying (Kol a) only as long as the given (Koln a) is not NULL, otherwise, evaluates to NULL.

forKoln :: (PgTyped a, PgTyped b) => Koln a -> (Kol a -> Kol b) -> Koln b

mapKoln with the arguments flipped.

bindKoln :: (PgTyped a, PgTyped b) => Koln a -> (Kol a -> Koln b) -> Koln b

Monadic bind like the one for Maybe.

Apply the given function to the underlying (Kol a) only as long as the given (Koln a) is not NULL, otherwise, evaluates to NULL.

altKoln :: PgTyped a => Koln a -> Koln a -> Koln a

Like (<|>) :: Maybe a -> Maybe a -> Maybe a.

Evaluates to the first argument if it is not NULL, otherwise evaluates to the second argument.

Querying

type Query = QueryArr ()

Query a is analogous to a Haskell value [a].

data QueryArr a b :: * -> * -> *

QueryArr a b is analogous to a Haskell function a -> [b].

leftJoin

Arguments

:: (Default Unpackspec a a, Default Unpackspec b b, Default NullMaker b nb)

=> Query a

-> Query b

-> ((a, b) -> Kol PGBool)

-> Query (a, nb)

Like opaleye's leftJoin, but the predicate is expected to return a Kol PGBool.

leftJoinExplicit

Arguments

:: Unpackspec a a

-> Unpackspec b b

-> NullMaker b nb

-> Query a

-> Query b

-> ((a, b) -> Kol PGBool)

-> Query (a, nb)

Like Opaleye's leftJoinExplicit, but the predicate is expected to return a Kol PGBool.

restrict :: QueryArr (Kol PGBool) ()

Like opaleye's restric, but takes a Kol as input.

nullTrue :: Arrow f => f (Koln PGBool) (Kol PGBool)

Convert a Koln PGBool to a (Kol PGBool). An outer NULL is converted to TRUE.

This can be used as a function or as a QueryArr, whatever works best for you. The QueryArr support is often convenient when working with restrict:

restrict <<< nullTrue -< ...

Simplified types:

nullTrue :: Koln PGBool -> Kol PGBool
nullTrue :: QueryArr (Koln PGBool) (Kol PGBool)

nullFalse :: Arrow f => f (Koln PGBool) (Kol PGBool)

Like nullTrue, but an outer NULL is converted to FALSE.

Booleans

lnot :: Op1' kol => kol PGBool -> kol PGBool

Logical NOT.

Note: This function can take any of Kol and Koln argument, with the return type being fully determined by it. The valid combinations are:

lnot :: Kol  PGBool -> Kol  PGBool
lnot :: Koln PGBool -> Koln PGBool

lor :: Op2 kol kol' kol'' => kol PGBool -> kol' PGBool -> kol'' PGBool

Logical OR. See eq for possible argument types.

land :: Op2 kol kol' kol'' => kol PGBool -> kol' PGBool -> kol'' PGBool

matchBool :: (Op3 kol kol' kol'' kol''', PgTyped a) => kol a -> kol' a -> kol'' PGBool -> kol''' a

Like bool, matchBool f t x evaluates to f if x is false, otherwise it evaluates to t.

Equality

eq :: (PgTyped a, Op2 kol kol' kol'') => kol a -> kol' a -> kol'' PGBool

Whether two column values are equal.

Note: This function can take any combination of Kol and Koln arguments, with the return type being fully determined by them. The valid combinations are:

eq :: Kol  x -> Kol  x -> Kol  PGBool
eq :: Kol  x -> Koln x -> Koln PGBool
eq :: Koln x -> Kol  x -> Koln PGBool
eq :: Koln x -> Koln x -> Koln PGBool

Mnemonic reminder: EQual.

member

Arguments

:: (PgTyped a, Op2 kol kol' kol', Op2' kol', FromKol kol', Foldable f)

=> kol a

-> f (kol' a)

-> kol' PGBool

Whether the given value is a member of the given collection.

Notice that a combination eq and or is more general that member, as they don't restrict kol'.

Comparisons

lt :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is less than the second. See eq for possible argument types.

Mnemonic reminder: Less Than.

lte :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is less than or equal to the second. See eq for possible argument types.

Mnemonic reminder: Less Than or Equal.

gt :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is greater than the second. See eq for possible argument types.

Mnemonic reminder: Greater Than.

gte :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is greater than or equal to the second. See eq for possible argument types.

Mnemonic reminder: Greater Than or Equal.

Selecting

col :: forall t c xs xs' a a'. HLensCxt (TC t c) Record xs xs' a a' => C c -> Lens (Rec t xs) (Rec t xs') a a'

Lens to a column.

Mnemonic: The COLumn.

Ordering

orderBy :: Order a -> Query a -> Query a

Order the rows of a Query according to the Order.

import Data.Monoid (<>)

-- Order by the first column ascending.  When first columns are equal
-- order by second column descending.
example :: Query (Column PGInt4, Column PGText)
        -> Query (Column PGInt4, Column PGText)
example = orderBy (asc fst <> desc snd)

asc :: PgOrd b => (a -> Kol b) -> Order a

Ascending order, no NULLs involved.

ascnf :: PgOrd b => (a -> Koln b) -> Order a

Ascending order, NULLs first.

ascnl :: PgOrd b => (a -> Koln b) -> Order a

Ascending order, NULLs last.

desc :: PgOrd b => (a -> Kol b) -> Order a

Descending order, no NULLs involved.

descnf :: PgOrd b => (a -> Koln b) -> Order a

Descending order, NULLs first.

descnl :: PgOrd b => (a -> Koln b) -> Order a

Descending order, NULLs last.

WDef

data WDef a

Whether to write a DEFAUT value or a specific value into a database column.

WDef is isomorphic to Maybe. It exists mainly to avoid accidentally mixing the two of them together.

Constructors

WDef
Write DEFAULT.

WVal a
Write the specified value.

wdef :: b -> (a -> b) -> WDef a -> b

Case analysis for WDef.

Like maybe, evaluates to the first argument if WDef, otherwise applies the given function to the a in WVal.

Types

data Col name wd rn pgType hsType

Column description.

This is only used as a promoted datatype expected to have kind Col Symbol WD RN * *.

name: Column name.

wd: Whether DEFAULT can be written to this column (WD) or not (W).

rn: Whether NULL might be read from this column (RN) or not (R).

pgType: Type of the column value used in Opaleye queries as index to Kol or Koln. This must be an instance of PgTyped.

hsType: Type of the column value used in Haskell outside Opaleye queries. Hint: don't use something like Maybe Bool here if you want to indicate that this is an optional Bool column. Instead, use Int here and RN in the rn field.

Notice that Col is very different from Kol and Koln: Kol and Koln are used at runtime for manipulating values stored in columns, Col is used to describe the properties of a column at compile time.

Constructors

Col name wd rn pgType hsType

data C c

Tag to be used alone or with Tagged for uniquely identifying a specific column in an unknown table.

Constructors

C

data T t

Tag to be used alone or with Tagged for uniquely identifying a specific table in a specific schema.

Constructors

Tabla t => T

data TC t c

Tag to be used alone or with Tagged for uniquely identifying a specific column in a specific table in a specific schema.

Constructors

Tabla t => TC

type TCa t c = Tagged (TC t c)

data RN

Whether to read a plain value or possibly NULL.

Constructors

R
Read plain value.

RN
Possibly read NULL.

data WD

Whether to write a specific value or possibly DEFAULT.

Constructors

W
Write a specific value.

WD
Possibly write DEFAULT. See WDef.

Individual columns

type Col_ByName t c = Col_ByName' c (Cols t)

Lookup a column in Tabla t by its name.

type family Col_Name col :: Symbol

Equations

Col_Name (Col n w r p h) = n

type family Col_PgRType col :: *

Equations

Col_PgRType (Col n w R p h) = Kol p

Col_PgRType (Col n w RN p h) = Koln p

type family Col_PgRNType col :: *

Equations

Col_PgRNType (Col n w r p h) = Koln p

type family Col_PgWType col :: *

Equations

Col_PgWType (Col n W r p h) = Col_PgRType (Col n W r p h)

Col_PgWType (Col n WD r p h) = WDef (Col_PgRType (Col n WD r p h))

type family Col_HsRType col :: *

Equations

Col_HsRType (Col n w R p h) = h

Col_HsRType (Col n w RN p h) = Maybe h

type family Col_HsIType col :: *

Equations

Col_HsIType (Col n W r p h) = Col_HsRType (Col n W r p h)

Col_HsIType (Col n WD r p h) = WDef (Col_HsRType (Col n WD r p h))

Set of columns

type Rec t xs = Tagged (T t) (Record xs)

All the representation of t used within opaleye-sot are (Rec t).

type Cols_HsR t = Map (Col_HsRFieldSym1 t) (Cols t)

Payload for (HsR t)

type Cols_HsI t = Map (Col_HsIFieldSym1 t) (Cols t)

Payload for (HsI t)

type Cols_PgR t = Map (Col_PgRSym1 t) (Cols t)

Payload for (PgR t)

type Cols_PgRN t = Map (Col_PgRNSym1 t) (Cols t)

Payload for (PgRN t)

type Cols_PgW t = Map (Col_PgWSym1 t) (Cols t)

Type of the Record columns when inserting or updating a row. Also, payload for (PgW t).

Column types

type family NotNullable x :: Constraint

Hack to workaround the current represenation for nullable columns. See Koln.

Equations

NotNullable (Nullable x) = "NotNullable" ~ "NotNullable: expected `x` but got `Nullable x`"

NotNullable x = ()

class PgPrimType (PgType a) => PgTyped a

Only PgTyped instances are allowed as indexes to Kol and Koln.

Associated Types

type PgType a :: *

PgType a indicates the primitive PostgreSQL column type that will ultimately be used as the index to opaleye's Column. This could be a itself, in the case of primitive types such as PGInt4, or it could be something else.

The motivation for this is the same as for “newtype wrappers” in Haskell. Say in you have this newtype in Haskell:

newtype UserId = UserId Int32

Now UserId and Int32 are different types, even though they share the same underlying representation Int32. With PgTyped you can achieve something similar:

instance PgTyped UserId where
  type PgType UserId = PGInt4

With that in place, you won't be able to accidentally mistake Kol PGInt4 for Kol UserId values, yet you will be able to easily reuse most of the machinery available to Kol PGInt4, except:

You will need a QueryRunnerColumnDefault instance to fetch Kol UserId values from the database as UserId:

  instance QueryRunnerColumnDefault PGInt4 UserId
  

You might find queryRunnerColumnFromWrapped useful for simple cases like UserId.

Notice that the instance mentions PGInt4 directly, not our PgTyped UserId. This is fine, not much would be gained by making the difference between them at this point, as you still need to do the FromRow parsing and any inconsistencies will be uncovered there.

If you want to reuse the exising Num or Fractional instances for Kol PGInt4, you will need to explicitely ask for it by instantiating PgNum UserId and/or PgFractional UserId. Most likely you won't need this for cases such as UserId@, since you shouldn't be doing arithmetic with user identifiers anyway.

class (PgTyped a, PgTyped b) => KolCoerce a b

KolCoerce a b says that Kol a can be safely coerced to Kol b using kolCoerce.

type family KolLike a :: Constraint

Satisfied if a is one of Kol or Koln. We use this Constraint to restrict the instances that can exist for typeclasses like FromKol.

Equations

KolLike Kol = ()

KolLike Koln = ()

class (PgTyped a, PGNum (PgType a)) => PgNum a

A PgNum a instance gives you a Num (Kol a) instance for free.

class (PgTyped a, PgNum a, PGFractional (PgType a)) => PgFractional a

A PgFractional a instance gives you a Fractional (Kol a) instance for free.

class PGOrd a

Typeclass for Postgres types which support ordering operations.

data PGBool :: *

data PGBytea :: *

data PGCitext :: *

data PGDate :: *

data PGFloat4 :: *

data PGFloat8 :: *

data PGInt2 :: *

data PGInt4 :: *

data PGInt8 :: *

data PGJsonb :: *

data PGJson :: *

data PGNumeric :: *

data PGText :: *

data PGTimestamptz :: *

data PGTimestamp :: *

data PGTime :: *

data PGUuid :: *

Miscellaneous

op1 :: (Op1_ fa fb ga gb, PgTyped a, PgTyped b) => (fa a -> fb b) -> ga a -> gb b

Convert an unary function fa a -> fb b to an unary function ga a -> gb b, where fa, fb, and ga can be any combination of Kol or Koln, with gb fully determined by them.

op2 :: (Op2_ fa fb fc ga gb gc, PgTyped a, PgTyped b, PgTyped c) => (fa a -> fb b -> fc c) -> ga a -> gb b -> gc c

Convert a binary function fa a -> fb b -> fc c to a binary function ga a -> gb b -> gc c, where fa, fb, fc, ga and gb can be any combination of Kol or Koln, with gc fully determined by them.