Opaleye.SOT.Internal

Description

This is an internal module. You are very discouraged from using it directly.

Documentation

type family NotNullable x :: Constraint

Hack to workaround the current represenation for nullable columns. See Koln.

Equations

NotNullable (Nullable x) = "NotNullable" ~ "NotNullable: expected `x` but got `Nullable x`"

NotNullable x = ()

class NotNullable a => PgPrimType a

Only PgPrimType instances are allowed as indexes to opaleye's Column.

You probably won't be adding new PgPrimType instances yourself, unless you are trying to represent a concrete PostgreSQL data type, but even then you might get away with creating PgTyped instances instead.

class PgPrimType (PgType a) => PgTyped a

Only PgTyped instances are allowed as indexes to Kol and Koln.

Associated Types

type PgType a :: *

PgType a indicates the primitive PostgreSQL column type that will ultimately be used as the index to opaleye's Column. This could be a itself, in the case of primitive types such as PGInt4, or it could be something else.

The motivation for this is the same as for “newtype wrappers” in Haskell. Say in you have this newtype in Haskell:

newtype UserId = UserId Int32

Now UserId and Int32 are different types, even though they share the same underlying representation Int32. With PgTyped you can achieve something similar:

instance PgTyped UserId where
  type PgType UserId = PGInt4

With that in place, you won't be able to accidentally mistake Kol PGInt4 for Kol UserId values, yet you will be able to easily reuse most of the machinery available to Kol PGInt4, except:

You will need a QueryRunnerColumnDefault instance to fetch Kol UserId values from the database as UserId:

  instance QueryRunnerColumnDefault PGInt4 UserId
  

You might find queryRunnerColumnFromWrapped useful for simple cases like UserId.

Notice that the instance mentions PGInt4 directly, not our PgTyped UserId. This is fine, not much would be gained by making the difference between them at this point, as you still need to do the FromRow parsing and any inconsistencies will be uncovered there.

If you want to reuse the exising Num or Fractional instances for Kol PGInt4, you will need to explicitely ask for it by instantiating PgNum UserId and/or PgFractional UserId. Most likely you won't need this for cases such as UserId@, since you shouldn't be doing arithmetic with user identifiers anyway.

class (PgTyped a, PGOrd (PgType a)) => PgOrd a

Like opaleye's PGOrd, but applies to a PgTyped.

class (PgTyped a, PGNum (PgType a)) => PgNum a

A PgNum a instance gives you a Num (Kol a) instance for free.

class (PgTyped a, PgNum a, PGFractional (PgType a)) => PgFractional a

A PgFractional a instance gives you a Fractional (Kol a) instance for free.

data Kol a

Like opaleye's (Column x), but with x guaranteed to be not Nullable. If you need to have a Nullable column type, use Koln instead.

Build using kol or Kol.

Notice that Kol is very different from Col: Col is used to describe the properties of a column at compile time. Kol is used at runtime for manipulating with values stored in columns.

We do not use (Column x), instead we use (Kol y) where x ~ PgType y. This is where we drift a bit appart from Opaleye. See https://github.com/tomjaguarpaw/haskell-opaleye/issues/97

Constructors

PgTyped a => Kol

Fields

unKol :: Column (PgType a)

liftKol1

Arguments

:: (PgTyped a, PgTyped b, Op1 kol kol')

=> (Column (PgType a) -> Column (PgType b))

-> kol a -> kol' b

Converts an unary function on Opaleye's Column to an unary function taking any of Kol and Koln as argument, with the result type fully determined by it.

liftKol2 :: (PgTyped a, PgTyped b, PgTyped c, Op2 kol kol' kol'') => (Column (PgType a) -> Column (PgType b) -> Column (PgType c)) -> kol a -> kol' b -> kol'' c

Converts a binary function on Opaleye's Columns to an binary function taking any of Kol and Koln as arguments, with the result type fully determined by them.

liftKol3 :: (PgTyped a, PgTyped b, PgTyped c, PgTyped d, Op3 kol kol' kol'' kol''') => (Column (PgType a) -> Column (PgType b) -> Column (PgType c) -> Column (PgType d)) -> kol a -> kol' b -> kol'' c -> kol''' d

Converts a ternary function on Opaleye's Columns to an ternary function on Kol.

Hint: You can further compose the result of this function with op3 to widen the range of accepted argument types.

class PgPrimType p => ToKol a p where

Build a Kol.

You need to provide a ToKol instance for every Haskell type you plan to convert to its PostgreSQL representation as Kol.

A a default implementation of kol is available for Wrapped instances:

default kol :: (PgTyped b, PgType b ~ p, Wrapped a, ToKol (Unwrapped a) p) => a -> Kol b
kol = kol . view _Wrapped'

Implementation notice: This class overlaps in purpose with Opaleye's Constant. Technicaly, we don't really need to repeat those instances here: we could just rely on Opaleye's Constant. However, as of today, Opaleye's Constant provides some undesired support which we deliberately want to avoid here. Namely, we don't want to support converting Int to PGInt4. If this is fixed upstream, we might go back to relying on Constant if suitable. See https://github.com/tomjaguarpaw/haskell-opaleye/pull/110

Minimal complete definition

Nothing

Methods

kol :: (PgTyped b, PgType b ~ p) => a -> Kol (b :: kb)

Convert a constant Haskell value (say, a Bool) to its equivalent PostgreSQL representation as a (Kol PGBool).

Some example simplified types:

kol :: Bool -> Kol PGBool
kol :: Int32 -> Kol PGInt4

data Koln a

Like opaleye's Column (Nullable x), but with x guaranteed to be not-Nullable.

Think of Koln a as Maybe (Kol a), with nul being analogous to the Nothing constructor and koln being analogous to the Just constructor.

Build safely using nul, koln, fromKol or Koln.

Notice that Koln is very different from Col: Col is used to describe the properties of a column at compile time. Koln is used at runtime for manipulating with values stored in columns.

We do not use Column (Nullable x), instead we use Koln y where x ~ PgType y. This is where we drift a bit appart from Opaleye. See https://github.com/tomjaguarpaw/haskell-opaleye/issues/97

Constructors

PgTyped a => Koln

Fields

unKoln :: Column (Nullable (PgType a))

koln :: (ToKol a p, PgTyped b, PgType b ~ p) => a -> Koln b

Build a Koln from a Haskell term. This is like the Just constructor for Maybe

nul :: PgTyped a => Koln a

PostgreSQL's NULL value. This is like the Nothing constructor for Maybe

fromKol :: PgTyped a => Kol a -> Koln a

Convert a Kol to a Koln.

matchKoln :: (PgTyped a, PgTyped b) => Kol b -> (Kol a -> Kol b) -> Koln a -> Kol b

Case analysis for Koln. Like maybe for Maybe.

If (Koln a) is NULL, then evaluate to the first argument, otherwise it applies the given function to the underlying (Kol a).

mapKoln :: (PgTyped a, PgTyped b) => (Kol a -> Kol b) -> Koln a -> Koln b

Like fmap for Maybe.

Apply the given function to the underlying (Kol a) only as long as the given (Koln a) is not NULL, otherwise, evaluates to NULL.

forKoln :: (PgTyped a, PgTyped b) => Koln a -> (Kol a -> Kol b) -> Koln b

mapKoln with the arguments flipped.

bindKoln :: (PgTyped a, PgTyped b) => Koln a -> (Kol a -> Koln b) -> Koln b

Monadic bind like the one for Maybe.

Apply the given function to the underlying (Kol a) only as long as the given (Koln a) is not NULL, otherwise, evaluates to NULL.

altKoln :: PgTyped a => Koln a -> Koln a -> Koln a

Like (<|>) :: Maybe a -> Maybe a -> Maybe a.

Evaluates to the first argument if it is not NULL, otherwise evaluates to the second argument.

class (PgTyped a, PgTyped b) => KolCoerce a b

KolCoerce a b says that Kol a can be safely coerced to Kol b using kolCoerce.

kolCoerce :: KolCoerce a b => Kol a -> Kol b

unsafeCoerceKol :: (PgTyped a, PgTyped b) => Kol a -> Kol b

type family KolLike a :: Constraint

Satisfied if a is one of Kol or Koln. We use this Constraint to restrict the instances that can exist for typeclasses like FromKol.

Equations

KolLike Kol = ()

KolLike Koln = ()

class KolLike f => FromKol f where

Methods

fromKol' :: PgTyped a => Kol a -> f a

Like fromKol, but generalizing the return type to either Kol or Koln.

data RN

Whether to read a plain value or possibly NULL.

Constructors

R
Read plain value.

RN
Possibly read NULL.

data WD

Whether to write a specific value or possibly DEFAULT.

Constructors

W
Write a specific value.

WD
Possibly write DEFAULT. See WDef.

data WDef a

Whether to write a DEFAUT value or a specific value into a database column.

WDef is isomorphic to Maybe. It exists mainly to avoid accidentally mixing the two of them together.

Constructors

WDef
Write DEFAULT.

WVal a
Write the specified value.

wdef :: b -> (a -> b) -> WDef a -> b

Case analysis for WDef.

Like maybe, evaluates to the first argument if WDef, otherwise applies the given function to the a in WVal.

data Col name wd rn pgType hsType

Column description.

This is only used as a promoted datatype expected to have kind Col Symbol WD RN * *.

name: Column name.

wd: Whether DEFAULT can be written to this column (WD) or not (W).

rn: Whether NULL might be read from this column (RN) or not (R).

pgType: Type of the column value used in Opaleye queries as index to Kol or Koln. This must be an instance of PgTyped.

hsType: Type of the column value used in Haskell outside Opaleye queries. Hint: don't use something like Maybe Bool here if you want to indicate that this is an optional Bool column. Instead, use Int here and RN in the rn field.

Notice that Col is very different from Kol and Koln: Kol and Koln are used at runtime for manipulating values stored in columns, Col is used to describe the properties of a column at compile time.

Constructors

Col name wd rn pgType hsType

type family Col_Name col :: Symbol

Equations

Col_Name (Col n w r p h) = n

data Col_NameSym0 col

type family Col_PgType col :: *

Equations

Col_PgType (Col n w r p h) = p

data Col_PgTypeSym0 col

type family Col_PgRType col :: *

Equations

Col_PgRType (Col n w R p h) = Kol p

Col_PgRType (Col n w RN p h) = Koln p

type family Col_PgRNType col :: *

Equations

Col_PgRNType (Col n w r p h) = Koln p

type family Col_PgWType col :: *

Equations

Col_PgWType (Col n W r p h) = Col_PgRType (Col n W r p h)

Col_PgWType (Col n WD r p h) = WDef (Col_PgRType (Col n WD r p h))

type family Col_HsRType col :: *

Equations

Col_HsRType (Col n w R p h) = h

Col_HsRType (Col n w RN p h) = Maybe h

type family Col_HsIType col :: *

Equations

Col_HsIType (Col n W r p h) = Col_HsRType (Col n W r p h)

Col_HsIType (Col n WD r p h) = WDef (Col_HsRType (Col n WD r p h))

type Col_ByName t c = Col_ByName' c (Cols t)

Lookup a column in Tabla t by its name.

type family Col_ByName' name cols :: Col Symbol WD RN * *

Equations

Col_ByName' n (Col n w r p h : xs) = Col n w r p h

Col_ByName' n (Col n' w r p h : xs) = Col_ByName' n xs

type HasColName t c = HasColName' c (Cols t)

type family HasColName' name cols :: Constraint

Equations

HasColName' n (Col n w r p h : xs) = ()

HasColName' n (Col n' w r p h : xs) = HasColName' n xs

type Cols_HsR t = Map (Col_HsRFieldSym1 t) (Cols t)

Payload for (HsR t)

type Col_HsRField t col = TCa t (Col_Name col) (Col_HsRType col)

data Col_HsRFieldSym1 t col

type Cols_HsI t = Map (Col_HsIFieldSym1 t) (Cols t)

Payload for (HsI t)

type Col_HsIField t col = TCa t (Col_Name col) (Col_HsIType col)

data Col_HsIFieldSym1 t col

type Cols_PgR t = Map (Col_PgRSym1 t) (Cols t)

Payload for (PgR t)

type family Col_PgR t col :: *

Equations

Col_PgR t (Col n w r p h) = TCa t n (Col_PgRType (Col n w r p h))

data Col_PgRSym1 t col

type Cols_PgRN t = Map (Col_PgRNSym1 t) (Cols t)

Payload for (PgRN t)

type family Col_PgRN t col :: *

Equations

Col_PgRN t (Col n w r p h) = TCa t n (Col_PgRNType (Col n w r p h))

data Col_PgRNSym1 t col

type Cols_PgW t = Map (Col_PgWSym1 t) (Cols t)

Type of the Record columns when inserting or updating a row. Also, payload for (PgW t).

type family Col_PgW t col :: *

Equations

Col_PgW t (Col n w r p h) = TCa t n (Col_PgWType (Col n w r p h))

data Col_PgWSym1 t col

data T t

Tag to be used alone or with Tagged for uniquely identifying a specific table in a specific schema.

Constructors

Tabla t => T

data TC t c

Tag to be used alone or with Tagged for uniquely identifying a specific column in a specific table in a specific schema.

Constructors

Tabla t => TC

type TCa t c = Tagged (TC t c)

data C c

Tag to be used alone or with Tagged for uniquely identifying a specific column in an unknown table.

Constructors

C

type Rec t xs = Tagged (T t) (Record xs)

All the representation of t used within opaleye-sot are (Rec t).

type HsR t = Rec t (Cols_HsR t)

Expected output type for runQuery on a (PgR t).

Important: If you are expecting a (PgR t) on the right side of a leftJoin, you will need to use (Maybe (PgR t)).

Mnemonic: Haskell Read.

type HsI t = Rec t (Cols_HsI t)

HsI t is the Haskell representation of Haskell values to be inserted to the database, as taken by "Opaleye.SOT.Run.runInsertTabla".

An HsI t can always be converted to a PgW t using pgWfromHsI, in case you need that for with the more general "Opaleye.SOT.Run.runInsert".

Mnemonic: Haskell Insert.

type PgR t = Rec t (Cols_PgR t)

Output type of queryTabla (T t).

Mnemonic: PostGresql Read.

type PgRN t = Rec t (Cols_PgRN t)

Like (PgRN t) but every field is Koln, as in the output type of the right hand side of a leftJoin with '(table t).

Mnemonic: PostGresql Read Nulls.

type PgW t = Rec t (Cols_PgW t)

Representation of PostgreSQL values to be written to the database. This type can be used as input for "Opaleye.SOT.Run.runInsert" and similar.

An HsI t can always be converted to a PgW t using pgWfromHsI, in

Mnemonic: PostGresql Write.

type ITabla t = (KnownSymbol (SchemaName t), KnownSymbol (TableName t), All PgTyped (Map Col_PgTypeSym0 (Cols t)), HDistributeProxy (Cols t), HMapAux HList (FnCol_Props t) (Map ProxySym0 (Cols t)) (Cols_Props t), HMapAux HList (HFmap FnPgWfromHsIField) (Cols_HsI t) (Cols_PgW t), HMapAux HList (HFmap FnPgWfromPgRField) (Cols_PgR t) (Cols_PgW t), HRLabelSet (Cols_HsR t), HRLabelSet (Cols_HsI t), HRLabelSet (Cols_PgR t), HRLabelSet (Cols_PgRN t), HRLabelSet (Cols_PgW t), SameLength (Cols_Props t) (Map ProxySym0 (Cols t)), SameLength (Cols_HsI t) (Cols_PgW t), SameLength (Cols_PgR t) (Cols_PgW t), ProductProfunctorAdaptor TableProperties (Record (Cols_Props t)) (Record (Cols_PgW t)) (Record (Cols_PgR t)), Default ColumnMaker (PgR t) (PgR t))

All these constraints need to be satisfied by tools that work with Tabla. It's easier to just write all the constraints once here and make ITabla a superclass of Tabla. Moreover, they enforce some sanity constraints on our Tabla so that we can get early compile time errors.

class ITabla t => Tabla t

Tabla means table in spanish.

An instance of this class can uniquely describe a PostgreSQL table and how to convert back and forth between it and its Haskell representation used when writing Opaleye queries.

The t type is only used as a tag for the purposes of uniquely identifying this Tabla.

Associated Types

type Database t :: *

Some kind of unique identifier used for telling appart the database where this table exists from other databases, so as to avoid accidentally mixing tables from different databases in queries.

type SchemaName t :: Symbol

PostgreSQL schema name where to find the table (defaults to "public", PostgreSQL's default schema name).

type TableName t :: Symbol

Table name.

type Cols t :: [Col Symbol WD RN * *]

Columns in this table. See the documentation for Col.

mkHsI

Arguments

:: (Tabla t, HRearrange (LabelsOf (Cols_HsI t)) xs (Cols_HsI t))

=> T t

-> ((forall c a. C c -> a -> TCa t c a) -> HList xs)

-> HsI t

Helper function to safely build an HsI t.

The type of this function isn't easy to understand, but an example should clarify its usage. Here we asume we have a Person datatype and a TPerson which is an instance of Tisch.

personToHsI :: Person -> HsI TPerson
personToHsI (Person name age) =
   mkHsI (T :: TPerson) $ \set_ -> hBuild
       (set_ (C :: C "name") name)
       (set_ (C :: C "age") age)

You are not required to use this function to build an HsI t if working with the tools from Data.HList is sufficient for you. This is just a convenience.

data FnPgWfromHsIField

Use with ApplyAB to apply convert a field in a (HList (Cols_HsI t)) to a field in a (HList (Cols_PgW t)).

Constructors

FnPgWfromHsIField

pgWfromHsI :: Tabla t => HsI t -> PgW t

Convert a custom Haskell type to a representation appropiate for inserting it as a new row using runInsert.

data FnPgWfromPgRField

Use with ApplyAB to apply convert a field in a (HList (Cols_PgR t)) to a field in a (HList (Cols_PgW t)).

Constructors

FnPgWfromPgRField

pgWfromPgR :: Tabla t => PgR t -> PgW t

Convert a (PgR t) resulting from a queryTable-like operation to a (PgW t) that can be used in a runUpdate-like operation.

colProps_wr :: PgTyped a => String -> TableProperties (Kol a) (Kol a)

Column properties: Write (no default), Read (not nullable).

colProps_wrn :: PgTyped a => String -> TableProperties (Koln a) (Koln a)

Column properties: Write (no default), Read (nullable).

colProps_wdr :: PgTyped a => String -> TableProperties (WDef (Kol a)) (Kol a)

Column properties: Write (optional default), Read (not nullable).

colProps_wdrn :: PgTyped a => String -> TableProperties (WDef (Koln a)) (Koln a)

Column properties: Write (optional default), Read (nullable).

type Cols_Props t = Map (Col_PropsSym1 t) (Cols t)

TableProperties for all the columns in Tabla t.

type Col_Props t col = TableProperties (Col_PgW t col) (Col_PgR t col)

TableProperties for a single column in Tabla t.

data Col_PropsSym1 t col

data Col_PropsSym0 col

class ICol_Props col where

Methods

colProps :: Tabla t => Proxy t -> Proxy col -> Col_Props t col

data FnCol_Props t

Use with ApplyAB to apply colProps to each element of an HList.

Constructors

FnCol_Props

table :: Tabla t => T t -> Table (PgW t) (PgR t)

Build the Opaleye Table for a Tabla.

queryTabla :: Tabla t => T t -> Query (PgR t)

Like opaleye's queryTable, but for a Tabla.

col :: forall t c xs xs' a a'. HLensCxt (TC t c) Record xs xs' a a' => C c -> Lens (Rec t xs) (Rec t xs') a a'

Lens to a column.

Mnemonic: The COLumn.

matchBool :: (Op3 kol kol' kol'' kol''', PgTyped a) => kol a -> kol' a -> kol'' PGBool -> kol''' a

Like bool, matchBool f t x evaluates to f if x is false, otherwise it evaluates to t.

lnot :: Op1' kol => kol PGBool -> kol PGBool

Logical NOT.

Note: This function can take any of Kol and Koln argument, with the return type being fully determined by it. The valid combinations are:

lnot :: Kol  PGBool -> Kol  PGBool
lnot :: Koln PGBool -> Koln PGBool

lor :: Op2 kol kol' kol'' => kol PGBool -> kol' PGBool -> kol'' PGBool

Logical OR. See eq for possible argument types.

lors :: (Op2' kol, FromKol kol, Foldable f) => f (kol PGBool) -> kol PGBool

Whether any of the given PGBools is true.

Notice that lor is more general that lors, as it doesn't restrict kol.

Mnemonic reminder: Logical ORs.

land :: Op2 kol kol' kol'' => kol PGBool -> kol' PGBool -> kol'' PGBool

lands :: (Op2' kol, FromKol kol, Foldable f) => f (kol PGBool) -> kol PGBool

Whether all of the given PGBools are true.

Notice that land is more general that lands, as it doesn't restrict kol.

Mnemonic reminder: Logical ANDs.

eq :: (PgTyped a, Op2 kol kol' kol'') => kol a -> kol' a -> kol'' PGBool

Whether two column values are equal.

Note: This function can take any combination of Kol and Koln arguments, with the return type being fully determined by them. The valid combinations are:

eq :: Kol  x -> Kol  x -> Kol  PGBool
eq :: Kol  x -> Koln x -> Koln PGBool
eq :: Koln x -> Kol  x -> Koln PGBool
eq :: Koln x -> Koln x -> Koln PGBool

Mnemonic reminder: EQual.

member

Arguments

:: (PgTyped a, Op2 kol kol' kol', Op2' kol', FromKol kol', Foldable f)

=> kol a

-> f (kol' a)

-> kol' PGBool

Whether the given value is a member of the given collection.

Notice that a combination eq and or is more general that member, as they don't restrict kol'.

lt :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is less than the second. See eq for possible argument types.

Mnemonic reminder: Less Than.

lte :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is less than or equal to the second. See eq for possible argument types.

Mnemonic reminder: Less Than or Equal.

gt :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is greater than the second. See eq for possible argument types.

Mnemonic reminder: Greater Than.

gte :: (Op2 kol kol' kol'', PgOrd a) => kol a -> kol' a -> kol'' PGBool

Whether the first argument is greater than or equal to the second. See eq for possible argument types.

Mnemonic reminder: Greater Than or Equal.

isNull :: PgTyped a => Koln a -> Kol PGBool

Whether a Koln is nul (NULL).

nullTrue :: Arrow f => f (Koln PGBool) (Kol PGBool)

Convert a Koln PGBool to a (Kol PGBool). An outer NULL is converted to TRUE.

This can be used as a function or as a QueryArr, whatever works best for you. The QueryArr support is often convenient when working with restrict:

restrict <<< nullTrue -< ...

Simplified types:

nullTrue :: Koln PGBool -> Kol PGBool
nullTrue :: QueryArr (Koln PGBool) (Kol PGBool)

nullFalse :: Arrow f => f (Koln PGBool) (Kol PGBool)

Like nullTrue, but an outer NULL is converted to FALSE.

restrict :: QueryArr (Kol PGBool) ()

Like opaleye's restric, but takes a Kol as input.

leftJoin

Arguments

:: (Default Unpackspec a a, Default Unpackspec b b, Default NullMaker b nb)

=> Query a

-> Query b

-> ((a, b) -> Kol PGBool)

-> Query (a, nb)

Like opaleye's leftJoin, but the predicate is expected to return a Kol PGBool.

leftJoinExplicit

Arguments

:: Unpackspec a a

-> Unpackspec b b

-> NullMaker b nb

-> Query a

-> Query b

-> ((a, b) -> Kol PGBool)

-> Query (a, nb)

Like Opaleye's leftJoinExplicit, but the predicate is expected to return a Kol PGBool.

asc :: PgOrd b => (a -> Kol b) -> Order a

Ascending order, no NULLs involved.

ascnl :: PgOrd b => (a -> Koln b) -> Order a

Ascending order, NULLs last.

ascnf :: PgOrd b => (a -> Koln b) -> Order a

Ascending order, NULLs first.

desc :: PgOrd b => (a -> Kol b) -> Order a

Descending order, no NULLs involved.

descnf :: PgOrd b => (a -> Koln b) -> Order a

Descending order, NULLs first.

descnl :: PgOrd b => (a -> Koln b) -> Order a

Descending order, NULLs last.

ppaUnTagged :: Profunctor p => p a b -> p (Tagged ta a) (Tagged tb b)

class Profunctor p => ProductProfunctorAdaptor p l ra rb | p l -> ra rb, p ra rb -> l where

A generalization of product profunctor adaptors such as p1, p4, etc.

The functional dependencies make type inference easier, but also forbid some otherwise acceptable instances.

Methods

ppa :: l -> p ra rb

class Op1_ Kol Kol g g => Op1' g

class Op1_ Kol Kol ga gb => Op1 ga gb

class (KolLike fa, KolLike fb, KolLike ga, KolLike gb) => Op1_ fa fb ga gb | fa fb ga -> gb where

Instances of this class can be used to convert an unary function fa a -> fb b to an unary function ga a -> gb b, where fa, fb, and ga can be any combination of Kol or Koln, with gb fully determined by them, biasing towards Kol.

Thanks to the KolLike superclass and the functional dependency in this class, it is impossible to add new Op1_ instances. All the possible instances are already exported from this module.

Methods

op1 :: (PgTyped a, PgTyped b) => (fa a -> fb b) -> ga a -> gb b

Convert an unary function fa a -> fb b to an unary function ga a -> gb b, where fa, fb, and ga can be any combination of Kol or Koln, with gb fully determined by them.

class Op2_ Kol Kol Kol g g g => Op2' g

class Op2_ Kol Kol Kol ga gb gc => Op2 ga gb gc

class (KolLike fa, KolLike fb, KolLike fc) => Op2_ fa fb fc ga gb gc | fa fb fc ga gb -> gc where

Instances of this class can be used to convert a binary function fa a -> fb b -> fc c to a binary function ga a -> gb b -> gc c, where fa, fb, fc, ga and gb can be any combination of Kol or Koln, with gc fully determined by them, biasing towards Kol.

Thanks to the KolLike superclass and the functional dependency in this class, it is impossible to add new Op2_ instances. All the possible instances are already exported from this module.

Methods

op2 :: (PgTyped a, PgTyped b, PgTyped c) => (fa a -> fb b -> fc c) -> ga a -> gb b -> gc c

Convert a binary function fa a -> fb b -> fc c to a binary function ga a -> gb b -> gc c, where fa, fb, fc, ga and gb can be any combination of Kol or Koln, with gc fully determined by them.

class Op3_ Kol Kol Kol Kol g g g g => Op3' g

class Op3_ Kol Kol Kol Kol ga gb gc gd => Op3 ga gb gc gd

class (KolLike fa, KolLike fb, KolLike fc, KolLike fd) => Op3_ fa fb fc fd ga gb gc gd | fa fb fc fd ga gb gc -> gd where

Instances of this class can be used to convert a ternary function fa a -> fb b -> fc c -> fd d to a ternary function ga a -> gb b -> gc c -> gd d, where fa, fb, fc, fd, ga, gb and gc can be any combination of Kol or Koln, with gd fully determined by them, biasing towards Kol.

Thanks to the KolLike superclass and the functional dependency in this class, it is impossible to add new Op3_ instances. All the possible instances are already exported from this module.

Methods

op3 :: (PgTyped a, PgTyped b, PgTyped c, PgTyped d) => (fa a -> fb b -> fc c -> fd d) -> ga a -> gb b -> gc c -> gd d

Convert a ternary function fa a -> fb b -> fc c -> fd d to a ternary function ga a -> gb b -> gc c -> gd d, where fa, fb, fc, fd, ga, gb and gc can be any combination of Kol or Koln, with gd fully determined by them.

unsafeUnNullableColumn :: Column (Nullable a) -> Column a

pgFloat4 :: Float -> Column PGFloat4

pgFloat8 :: Float -> Column PGFloat8

type family All c xs :: Constraint

Apply a same constraint to all the types in the list.

Equations

All c `[]` = ()

All c (x : xs) = (c x, All c xs)

data ProxySym0 a

Defunctionalized Proxy. To be used with Apply.

class HDistributeProxy xs where

Methods

hDistributeProxy :: Proxy xs -> HList (Map ProxySym0 xs)