Opaleye.SOT.Run

Connection

data Conn perms

Conn perms is just a wrapper around postgresql-simple's Connection that carries, in perms, type-level information about which kind of operations can be performed on this connection.

The problem with Connection is that it allows us to perform any operation on it (reading, inserting, commiting a transaction, etc.), but sometimes we would prefer to limit the type of operations this connection might support. For example, we might want to forbid transactions from being commited because they are being handled by someone else. Conn perms carries type-level information about this.

Note that Conn is not thread-safe, you are encouraged to maintain a multithreaded pool of Conn instead. See Data.Pool from the ex-pool package.

type Conn' = Conn `[Fetch, Insert, Update, Delete, Transact]`

A type synonym for a Conn with all the permissions enabled.

unConn :: Conn ps -> Connection

connect :: MonadIO m => ConnectInfo -> m Conn'

Return a new connection.

connect' :: MonadIO m => ByteString -> m Conn'

Like connect, except it takes a libpq connection string instead of a ConnectInfo.

close :: (MonadIO m, MonadMask m) => Conn ps -> m ()

Warning: Using the given Conn ps after calling close will result in a runtime exception.

Permissions

data Perm

Perm is only used at the type level to index Conn.

Constructors

Fetch
Allow fetching a value from the database (i.e., SELECT, ... RETURNING).

Update
Allow updating data in the database (i.e., UPDATE)

Insert
Allow inserting new data to the database (i.e., INSERT)

Delete
Allow deleting data from the database (i.e., DELETE)

Transact
Allow starting and finishing transactions (i.e., BEGIN, COMMIT, ROLLBACK).

Savepoint
Allow creating transactions savepoints and rolling back to them.

type family Allow p ps :: Constraint

Allow p ps ensures that p is present in ps.

The kind of p can be Perm or [Perm].

Equations

Allow (`[]` :: [Perm]) ps = ()

Allow (p : ps :: [Perm]) qs = (Allow p qs, Allow ps qs)

Allow (p :: Perm) (p : ps) = ()

Allow (p :: Perm) (q : ps) = Allow p ps

type family Forbid p ps :: Constraint

Forbid' p ps ensures that p is not present in ps.

The kind of p can be Perm or [Perm].

Equations

Forbid (`[]` :: [Perm]) ps = ()

Forbid (p : ps :: [Perm]) qs = (Forbid p qs, Forbid ps qs)

Forbid (p :: Perm) (p : ps) = "Opaleye.SOT.Run.Forbid" ~ "Forbid: The forbidden permission is allowed"

Forbid (p :: Perm) (q : ps) = Forbid p ps

Forbid (p :: Perm) `[]` = ()

type family DropPerm p ps :: [Perm]

DropPerm p ps removes p from ps if present.

The kind of p can be Perm or [Perm].

Equations

DropPerm (`[]` :: [Perm]) ps = ps

DropPerm (p : ps :: [Perm]) qs = DropPerm p (DropPerm ps qs)

DropPerm (p :: Perm) (p : ps) = DropPerm p ps

DropPerm (p :: Perm) (q : ps) = q : DropPerm p ps

DropPerm (p :: Perm) `[]` = `[]`

withoutPerm

Arguments

:: (MonadIO m, MonadMask m, Allow p ps, ps' ~ DropPerm p ps)

=> proxy (p :: k)
k may be Perm or [Perm].

-> Conn ps

-> (Conn ps' -> m a)
The usage of Conn ps is undefined within this function, and Conn ps' mustn't escape the scope of this function.

-> m a

Drop a permission from the connection.

Transaction

data IsolationLevel

Like IsolationLevel, but without support for default values.

Constructors

ReadCommitted

RepeatableRead

Serializable

withTransactionRead

Arguments

:: (MonadIO m, MonadMask m, Allow Transact ps, Forbid Savepoint ps, ps' ~ DropPerm `[Transact, Insert, Update, Delete]` ps)

=> IsolationLevel

-> Conn ps

-> (Conn ps' -> m a)
The usage of Conn ps is undefined within this function, as well as the usage of Conn ps' outside this function.

-> m a

Execute the given callback within a read-only transaction with the given isolation level. The transaction is rolled-back afterwards, as there wouldn't be anything to commit anyway, even in case of execeptions.

withTransactionReadWrite

Arguments

:: (MonadIO m, MonadMask m, Allow Transact ps, Forbid Savepoint ps, ps' ~ (Savepoint : DropPerm Transact ps))

=> IsolationLevel

-> Conn ps

-> (Conn ps' -> m (Either a b))
The usage of Conn ps is undefined within this function, as well as the usage of Conn ps' outside this function. A Left return value rollbacks the transaction, Right commits it.

-> m (Either a b)

Execute the given callback within a read-write transaction with the given isolation level, rolling back the transaction in case of exceptions, and either commiting or rolling back the transaction otherwise, as requested by the passed in callback.

withSavepoint

Arguments

:: (MonadIO m, MonadMask m, Allow Savepoint ps, Forbid Transact ps)

=> Conn ps

-> (Conn ps -> m (Either a b))
A Left return value rollbacks the savepoint, Right keeps it.

-> m (Either a b)

You can use this function within withTransaction as a sort of nested transaction.

Query

runQuery

Arguments

:: (MonadIO m, MonadThrow m, Default QueryRunner v r, Allow Fetch ps)

=> Conn ps

-> Query v

-> m [r]

Query and fetch zero or more resulting rows.

runQuery1

Arguments

:: (MonadIO m, MonadThrow m, Default QueryRunner v r, Allow Fetch ps)

=> Conn ps

-> Query v

-> m (Maybe r)

Query and fetch zero or one resulting row.

Throws ErrNumRows if there is more than one row in the result.

Insert

runInsert

Arguments

:: (MonadIO m, MonadThrow m, Allow Insert ps, Foldable f)

=> Conn ps

-> Table w v

-> f w

-> m ()

Insert many rows.

Throws ErrNumRows if the number of actually affected rows is different than the number of passed in rows. Use runInsert' if you don't want this behavior.

runInsert'

Arguments

:: (MonadIO m, Allow Insert ps, Foldable f)

=> Conn ps

-> Table w v

-> f w

-> m Int64

Like runInsert, but doesn't check the number of affected rows. Returns the number of affected rows.

runInsert1

Arguments

:: (MonadIO m, MonadThrow m, Allow Insert ps)

=> Conn ps

-> Table w v

-> w

-> m ()

Insert one row.

Throws ErrNumRows if the number of actually affected rows is different than one. Use runInsert' if you don't want this behavior.

runInsertTabla

Arguments

:: (MonadIO m, MonadThrow m, Allow Insert ps, Tabla t, Foldable f)

=> Conn ps

-> T t

-> f (HsI t)

-> m ()

Like runInsert, but easier to use if you are querying a single Tabla.

runInsertTabla'

Arguments

:: (MonadIO m, MonadThrow m, Allow Insert ps, Tabla t, Foldable f)

=> Conn ps

-> T t

-> f (HsI t)

-> m Int64

Like runInsert', but easier to use if you are querying a single Tabla.

runInsertTabla1

Arguments

:: (MonadIO m, MonadThrow m, Allow Insert ps, Tabla t)

=> Conn ps

-> T t

-> HsI t

-> m ()

Like runInsert1, but easier to use if you are querying a single Tabla.

Returning

runInsertReturning

Arguments

:: (MonadIO m, MonadThrow m, Default QueryRunner v r, Allow `[Insert, Fetch]` ps, Foldable f)

=> Conn ps

-> Table w v

-> f w

-> m [r]

Insert many rows, returning data from the rows actually inserted.

Throws ErrNumRows if the number of actually affected rows is different than the number of passed in rows. Use runInsertReturning' if you don't want this behavior.

runInsertReturning'

Arguments

:: (MonadIO m, Default QueryRunner v r, Allow `[Insert, Fetch]` ps, Foldable f)

=> Conn ps

-> Table w v

-> f w

-> m [r]

runInsertReturning1

Arguments

:: (MonadIO m, MonadThrow m, Default QueryRunner v r, Allow `[Insert, Fetch]` ps)

=> Conn ps

-> Table w v

-> w

-> m r

Insert one row, returning data from the one row actually inserted.

Throws ErrNumRows if the number of actually affected rows is different than the one. Use runInsertReturning' if you don't want this behavior.

Update

runUpdate

Arguments

:: (MonadIO m, Allow Update ps)

=> Conn ps

-> Table w r

-> (r -> w)

-> (r -> Kol PGBool)

-> m Int64

Like opaleye's runUpdate, but the predicate is expected to return a Kol PGBool. Returns the number of affected rows.

It is recommended that you use runUpdateTabla if you are trying to update a table that is an instance of Tabla. The result is the same, but this function might be less convenient to use.

runUpdateTabla

Arguments

:: (Tabla t, MonadIO m, Allow Update ps)

=> Conn ps

-> T t

-> (PgW t -> PgW t)

-> (PgR t -> Kol PGBool)

-> m Int64

Like runUpdateTabla', but takes t explicitely for the times when it can't be inferred.

runUpdateTabla'

Arguments

:: (Tabla t, MonadIO m, Allow Update ps)

=> Conn ps

-> (PgW t -> PgW t)
Upgrade current values to new values.

-> (PgR t -> Kol PGBool)
Whether a row should be updated.

-> m Int64
Number of updated rows.

Like runUpdate, but specifically designed to work well with Tabla.

Delete

runDelete

Arguments

:: (MonadIO m, Allow Delete ps)

=> Conn ps

-> Table w r

-> (r -> Kol PGBool)

-> m Int64

Like opaleye's runDelete, but the predicate is expected to return a Kol PGBool. Returns the number of affected rows.

It is recommended that you use runDeleteTabla if you are trying to update a table that is an instance of Tabla, the result is the same, but this function might be less convenient to use.

runDeleteTabla

Arguments

:: (Tabla t, MonadIO m, Allow Delete ps)

=> Conn ps

-> T t

-> (PgR t -> Kol PGBool)
Whether a row should be deleted.

-> m Int64

Like runDeleteTabla', but takes t explicitely for the times when it can't be inferred.

runDeleteTabla'

Arguments

:: (Tabla t, MonadIO m, Allow Delete ps)

=> Conn ps

-> (PgR t -> Kol PGBool)
Whether a row should be deleted.

-> m Int64

Like runDelete, but specifically designed to work well with Tabla.

Exception

data ErrNumRows

Exception thrown when indicating less rows than expected are available.

Constructors

ErrNumRows

Fields

_errNumRowsExpected :: Int64

_errNumRowsActual :: Int64

_errNumRowsSQL :: Maybe String

Nothing means empty query.

Parsing results

class QueryRunnerColumnDefault pgType haskellType where

A QueryRunnerColumnDefault pgType haskellType represents the default way to turn a pgType result from the database into a Haskell value of type haskelType.

Methods

queryRunnerColumnDefault :: QueryRunnerColumn pgType haskellType

qrcFromField :: FromField b => QueryRunnerColumn a b

qrcFieldParser :: FieldParser b -> QueryRunnerColumn a b

qrcFieldParserMap :: QueryRunnerColumnDefault a b => (FieldParser b -> FieldParser b') -> QueryRunnerColumn a b'

qrcMap :: QueryRunnerColumnDefault a b => (b -> b') -> QueryRunnerColumn a b'

qrcMapMay :: (QueryRunnerColumnDefault a b, Show b, Typeable b') => (b -> Maybe b') -> QueryRunnerColumn a b'

qrcPrism :: (QueryRunnerColumnDefault a b, Show b, Typeable b') => Prism' b b' -> QueryRunnerColumn a b'

qrcWrapped :: (Wrapped b, QueryRunnerColumnDefault a (Unwrapped b)) => QueryRunnerColumn a b