Opaleye.SOT

Description

For a better experience, it is recommended that you import this module unqualified as follows:

import           Opaleye.SOT
import qualified Opaleye as O

This module overrides some of the names exported by the Opaleye.

This module doesn't export any infix operator.

Defining a Tabla

class ITabla t => Tabla t

Tabla means table in spanish.

An instance of this class can uniquely describe a PostgreSQL table and how to convert back and forth between it and its Haskell representation used when writing Opaleye queries.

The t type is only used as a tag for the purposes of uniquely identifying this Tabla.

Associated Types

type Database t :: *

Some kind of unique identifier used for telling appart the database where this table exists from other databases, so as to avoid accidentally mixing tables from different databases in queries.

type SchemaName t :: Symbol

PostgreSQL schema name where to find the table (defaults to "public", PostgreSQL's default schema name).

type TableName t :: Symbol

Table name.

type Cols t :: [Col Symbol WD RN * *]

Columns in this table. See the documentation for Col.

class Tabla t => UnHsR t a where

Convert an Opaleye-compatible Haskell representation of a to a when reading from the database.

Notice that you are not required to provide instances of this class if working with HsR t is sufficient for you, or if you already have a function (HsR t -> a) at hand. Nevertheless, readability wise, it can be useful to have a single overloaded function used to decode each (HsR t).

Methods

unHsR' :: HsR t -> Either SomeException a

Convert an Opaleye-compatible Haskell representation of a to a.

For your convenience, you are encouraged to use cola, but you may also use other tools from Data.HList.Record as you see fit:

unHsR' r = Person (r ^. cola (C :: C "name"))
                  (r ^. cola (C :: C "age"))

Hint: If the type checker is having trouble inferring (HsR t), consider using unHsR instead.

unHsR :: UnHsR t a => T t -> HsR t -> Either SomeException a

Like unHsR', except it takes t explicitely for the times when it can't be inferred.

unHsR_ :: UnHsR t a => T t -> Proxy a -> HsR t -> Either SomeException a

Like unHsR', except it takes both t and a explicitely for the times when they can't be inferred.

class Tabla t => ToHsI t a where

Build a (HsR t) representation for a for inserting it to the database.

Notice that you are not required to provide instances of this class if working with HsI t is sufficient for you, or if you already have a function (a -> HsI t) at hand.

Methods

toHsI' :: a -> HsI t

Convert an a to an Opaleye-compatible Haskell representation to be used when inserting a new row to this table.

For your convenience, you may use mkHsI together with hBuild to build toHsI:

toHsI (Person name age) = mkHsI $ \set_ -> hBuild
    (set_ (C :: C "name") name)
    (set_ (C :: C "age") age)

You may also use other tools from Data.HList.Record as you see fit.

Hint: If the type checker is having trouble inferring (HsI t), consider using toHsI instead. Nevertheless, it is more likely that you use toPgW directly, which skips the HsI intermediate representation altogether.

toHsI :: ToHsI t a => T t -> a -> HsI t

Like toHsI', except it takes t explicitely for the times when it can't be inferred.

mkHsI

Arguments

:: (Tabla t, HRearrange (LabelsOf (Cols_HsI t)) xs (Cols_HsI t))

=> ((forall c a. C c -> a -> Tagged (TC t c) a) -> HList xs)

-> HsI t

Convenience intended to be used within toHsI', together with hBuild.

toHsI (Person name age) = mkHsI $ \set_ -> hBuild
    (set_ (C :: C "name") name)
    (set_ (C :: C "age") age)

toPgW' :: ToHsI t a => a -> PgW t

Convert a custom Haskell type to a representation appropiate for inserting it as a new row.

toPgW :: ToHsI t a => T t -> a -> PgW t

Like toPgW', but takes an explicit t.

Working with Tabla

table' :: forall t. Tabla t => Table (PgW t) (PgR t)

Build the Opaleye Table for a Tabla.

table :: Tabla t => T t -> Table (PgW t) (PgR t)

Like table', but takes t explicitly to help the compiler when it can't infer t.

queryTabla' :: Tabla t => Query (PgR t)

Like Opaleye's queryTable, but for a Tabla.

queryTabla :: Tabla t => T t -> Query (PgR t)

Like queryTabla', but takes t explicitly to help the compiler when it can't infer t.

type HsR t = Rec t (Cols_HsR t)

Expected output type for runQuery on a (PgR t).

Important: If you are expecting a (PgR t) on the right side of a leftJoin, you will need to use (Maybe (PgR t)).

Mnemonic: Haskell Read.

type HsI t = Rec t (Cols_HsI t)

Output type of toHsI, used when inserting a new row to the table.

This type is used internally as an intermediate representation between your own Haskell representation for a to-be-inserted t and (PgW t).

Mnemonic: Haskell Insert.

type PgR t = Rec t (Cols_PgR t)

Output type of queryTabla (T t).

Mnemonic: PostGresql Read.

type PgRN t = Rec t (Cols_PgRN t)

Like (PgRN t) but every field is Koln, as in the output type of the right hand side of a leftJoin with '(table t).

Mnemonic: PostGresql Read Nulls.

type PgW t = Rec t (Cols_PgW t)

Representation of (ToHsI t) as Kols. To be used when writing to the database.

Mnemonic: PostGresql Write.

update' :: Tabla t => PgR t -> PgW t

Convert a (PgR t) resulting from a queryTable-like operation to a (PgW t) that can be used in a runUpdate-like operation.

update :: Tabla t => T t -> PgR t -> PgW t

Like update', but takes an explicit t for when it can't be inferred.

Kol

data Kol a

Like Opaleye's (Column a), but guaranteed to be not Nullable. If you need to have a Nullable column type, use Koln instead.

Build using kol, view using unKol.

We do not use (Column a), instead we use (Kol a) This is where we drift a bit appart from Opaleye, but hopefully not for a long time. See https://github.com/tomjaguarpaw/haskell-opaleye/issues/97

Notice that Kol is very different from Col: Col is used to describe the properties of a column at compile time. Kol is used at runtime for manipulating with values stored in columns.

unKol :: Kol a -> Column a

class ToKol hs pg where

Build a Kol.

You need to provide an instance for every Haskell type you plan to convert to its PostgreSQL representation as Kol.

A a default implementation of kol is available for Wrapped instances.

default kol :: (Wrapped hs, ToKol (Unwrapped hs) pg) => hs -> Kol pg
kol = kol . view _Wrapped'

Implementation notice: This class overlaps in purpose with Opaleye's Constant. Technicaly, we don't really need to repeat those instances here: we could just rely on Opaleye's Constant. However, as of today, Opaleye's Constant provides some undesired support which we deliberately want to avoid here. Namely, we don't want to support converting Int to PGInt4. As soon as that is fixed upstream, we might go back to relying on Constant if suitable. See https://github.com/tomjaguarpaw/haskell-opaleye/pull/110

Minimal complete definition

Nothing

Methods

kol :: hs -> Kol pg

Build a Kol.

You can use this function to convert an Opaleye Column to a Kol, or to convert a constant Haskell value (say, a Bool) to its equivalent PostgreSQL representation as a (Kol PGBool).

Some example simplified types:

kol :: PGType a => Column a -> Kol a
kol :: Bool -> Kol PGBool
kol :: Int32 -> Kol PGInt4

liftKol1 :: PGType b => (Column a -> Column b) -> Kol a -> Kol b

Converts an unary function on Opaleye's Column to an unary function on Kol.

Hint: You can further compose the result of this function with op1 to widen the range of accepted argument types.

liftKol2 :: PGType c => (Column a -> Column b -> Column c) -> Kol a -> Kol b -> Kol c

Converts a binary function on Opaleye's Columns to an binary function on Koln.

Hint: You can further compose the result of this function with op2 to widen the range of accepted argument types.

Koln

type family NotNullable x :: Constraint

Horrible hack to workaround the current represenation for nullable columns. See Koln.

Equations

NotNullable (Nullable x) = "NotNullable" ~ "NotNullable: expected `x` but got `Nullable x`"

NotNullable x = ()

data Koln a

Like (Column (Nullable a)), but the a is guaranteed to be not-Nullable.

Build safely using koln, view using unKoln.

We do not use (Column (Nullable a)), but instead we use (Koln a). This is where we drift a bit appart from Opaleye, but hopefully not for a long time. See https://github.com/tomjaguarpaw/haskell-opaleye/issues/97

Think of (Koln a) as (Maybe a), where (Nothing == nul) and (Just a == Column a).

Notice that Koln is very different from Col: Col is used to describe the properties of a column at compile time. Koln is used at runtime for manipulating with values stored in columns.

unKoln :: Koln a -> Column (Nullable a)

koln :: ToKoln hs pg => hs -> Koln pg

Build a Koln.

koln :: Kol a -> Koln a
koln :: PGType a => Column (Nullable a) -> Koln a
koln :: ToKol hs pg => hs -> Koln pg
koln :: ToKol hs pg => Maybe hs -> Koln pg -- NULL if Nothing

matchKoln :: Kol b -> (Kol a -> Kol b) -> Koln a -> Kol b

Like maybe. Case analysis for Koln.

If (Koln a) is NULL, then evaluate to the first argument, otherwise it applies the given function to the underlying (Kol a).

isNull :: GetKoln w a => w -> Kol PGBool

Like Opaleye's isNull, but works for any GetKoln.

mapKoln :: (Kol a -> Kol b) -> Koln a -> Koln b

Like fmap for Maybe.

Apply the given function to the underlying (Kol a) only as long as the given (Koln a) is not NULL, otherwise, evaluates to NULL.

bindKoln :: Koln a -> (Kol a -> Koln b) -> Koln b

Monadic bind like the one for Maybe.

Apply the given function to the underlying (Kol a) only as long as the given (Koln a) is not NULL, otherwise, evaluates to NULL.

altKoln :: Koln a -> Koln a -> Koln a

Like ((<|>) :: Maybe a -> Maybe a -> Maybe a).

Evaluates to the first argument if it is not NULL, otherwise evaluates to the second argument.

liftKoln1

Arguments

:: PGType b

=> (Column (Nullable a) -> Column (Nullable b))

-> Koln a -> Koln b

Converts an unary function on Opaleye's Nullable Column to an unary function on Koln.

Hint: You can further compose the result of this function with op1 to widen the range of accepted argument types.

liftKoln2

Arguments

:: PGType c

=> (Column (Nullable a) -> Column (Nullable b) -> Column (Nullable c))

-> Koln a -> Koln b -> Koln c

Converts a binary function on Opaleye's Nullable Columns to a binary function on Kolns.

Hint: You can further compose the result of this function with op2 to widen the range of accepted argument types.

Executing

module Opaleye.SOT.Run

Querying

leftJoin

Arguments

:: (Default Unpackspec a a, Default Unpackspec b b, Default NullMaker b nb, GetKol gkb PGBool)

=> Query a

-> Query b

-> ((a, b) -> gkb)

-> Query (a, nb)

Like Opaleye's leftJoin, but the predicate is expected to return a (GetKol w PGBool).

leftJoinExplicit

Arguments

:: GetKol gkb PGBool

=> Unpackspec a a

-> Unpackspec b b

-> NullMaker b nb

-> Query a

-> Query b

-> ((a, b) -> gkb)

-> Query (a, nb)

Like Opaleye's leftJoinExplicit, but the predicate is expected to return a (GetKol w PGBool).

restrict :: GetKol w PGBool => QueryArr w ()

Like Opaleye's restric, but takes a Kol as input.

restrict :: QueryArr (Kol PGBool) ()
restrict :: QueryArr (Kol (Tagged (TC t c) PGBool)) ()

nullTrue :: (Arrow f, GetKoln w PGBool) => f w (Kol PGBool)

Flatten (Koln PGBool) or compatible (see GetKoln) to (Kol PGBool). An outer NULL is converted to TRUE.

This can be used as a function or as a QueryArr, whatever works best for you. The QueryArr support is often convenient when working with restrict:

restrict <<< nullTrue -< ...

Simplified types:

nullTrue :: Koln PGBool -> Kol PGBool
nullTrue :: Koln (Tagged (TC t c) PGBool) -> Kol PGBool
nullTrue :: QueryArr (Koln PGBool) (Kol PGBool)
nullTrue :: QueryArr (Koln (Tagged (TC t c) PGBool)) (Kol PGBool)

nullFalse :: (Arrow f, GetKoln w PGBool) => f w (Kol PGBool)

Like nullTrue, but an outer NULL is converted to FALSE.

no :: Op_no a b => a -> b

Polymorphic Opaleye's not. See eq for the type of arguments this function can take.

“No” means “not” in English, Spanish, and Italian, and it is a great name because it doesn't clash with not.

type Op_no a b = Op1 PGBool PGBool (Kol PGBool) (Kol PGBool) a b

Constraint on arguments to no.

eq :: Op_eq x a b c => a -> b -> c

Polymorphic Opaleye's (.==).

Mnemonic reminder: EQual.

eq :: Kol x -> Kol x -> Kol PGBool
eq :: Kol x -> Koln x -> Koln PGBool
eq :: Kol x -> Tagged t (Kol x) -> Kol PGBool
eq :: Kol x -> Tagged t (Koln x) -> Koln PGBool
eq :: Koln x -> Kol x -> Koln PGBool
eq :: Koln x -> Koln x -> Koln PGBool
eq :: Koln x -> Tagged t (Kol x) -> Koln PGBool
eq :: Koln x -> Tagged t (Koln x) -> Koln PGBool

Any of the above combinations with the arguments fliped is accepted too. Additionally, a Comparable constraint will be required if you try to compare two Tabla-aware columns directly; that is, a Kol or a Koln wrapped in a (TC t c), such as those obtained with (view . col):

Simplified type signature just so that you get an idea:

eq :: Comparable t1 c1 t2 c2
   => Tagged (TC t1 c1) a
   -> Tagged (TC t2 c2) b
   -> Koln PGBool

Important: Opaleye's Column is deliberately not supported. Use kol or koln to convert a Column to a Kol or Koln respectively.

Debugging hint: If the combination of a and b that you give to eq is unacceptable, you will get an error from the typechecker saying that an Op2 instance is missing. Do not try to add a new instance for Op2, it is an internal class that already supports all the possible combinations of x, a, b, and c. Instead, make sure your are not trying to do something funny such as comparing two Kolns for equality and expecting a Kol as a result (that is, you would be trying to compare two nullable columns and ignoring the possibilty that one of the arguments might be NULL, leading to a NULL result).

type Op_eq x a b c = Op2 x x PGBool (Kol x) (Kol x) (Kol PGBool) a b c

Constraints on arguments to eq.

Given as a and b any combination of (Kol x), (Koln x) or their respective wrappings in (Tagged (TC t c)), get c as result, which will be (Koln PGBool) if there was a (Koln x) among the given arguments, otherwise it will be (Kol PGBool).

This type synonym is exported for two reasons:

It increases the readability of the type of eq and any type errors resulting from its misuse.

If you are taking any of a or b as arguments to a function where eq is used, then you will need to ensure that some constraints are satisfied by those arguments. Adding Op_eq as a constraint to that function will solve the problem.

To keep in mind: The type c is fully determined by x, a, and b. This has the practical implication that when both (Kol z) and (Koln z) would be suitable types for c, we make a choice and prefer to only support (Kol z), leaving you to use koln on the return type if you want to convert it to (Koln z). This little inconvenience, however, significantly improves type inference when using eq.

lt :: Op_lt x a b c => a -> b -> c

Like Opaleye's (.=<), but can accept more arguments than just Column. See eq for a detailed explanation.

Mnemonic reminder: Less Than.

type Op_lt x a b c = (PGOrd x, Op2 x x PGBool (Kol x) (Kol x) (Kol PGBool) a b c)

Constraint on arguments to lt. See Op_eq for a detailed explanation.

ou :: Op_ou a b c => a -> b -> c

Like Opaleye's (.||), but can accept more arguments than just Column. See eq for a detailed explanation.

“Ou” means “or” in French, and it is a great name because it doesn't overlap with or. N'est-ce pas?

type Op_ou a b c = Op2 PGBool PGBool PGBool (Kol PGBool) (Kol PGBool) (Kol PGBool) a b c

Constraint on arguments to ou. See Op_eq for a detailed explanation.

et :: Op_et a b c => a -> b -> c

Like Opaleye's (.&&), but can accept more arguments than just Column. See eq for a detailed explanation.

“Et” means “and” in French, and it is a great name because it doesn't overlap with and. N'est-ce pas?

type Op_et a b c = Op2 PGBool PGBool PGBool (Kol PGBool) (Kol PGBool) (Kol PGBool) a b c

Constraint on arguments to et. See Op_eq for a detailed explanation.

Selecting

col :: HLensCxt (TC t c) Record xs xs' a a' => C c -> Lens (Rec t xs) (Rec t xs') (Tagged (TC t c) a) (Tagged (TC t c) a')

Lens to the value of a column.

Mnemonic: the COLumn.

cola :: forall t c xs xs' a a'. HLensCxt (TC t c) Record xs xs' a a' => C c -> Lens (Rec t xs) (Rec t xs') a a'

Lens to the value of a column without the TC tag.

Most of the time you'll want to use col instead, but this might be more useful when trying to change the type of a during an update, or when implementing unHsR.

Mnemonic: The COLumn's A.

Ordering

asc :: (GetKol w b, PGOrd b) => (a -> w) -> Order a

Ascending order, no NULLs involved.

ascnf :: (GetKoln w b, PGOrd b) => (a -> w) -> Order a

Ascending order, NULLs first.

ascnl :: (GetKoln w b, PGOrd b) => (a -> w) -> Order a

Ascending order, NULLs last.

desc :: (GetKol w b, PGOrd b) => (a -> w) -> Order a

Descending order, no NULLs involved.

descnf :: (GetKoln w b, PGOrd b) => (a -> w) -> Order a

Descending order, NULLs first.

descnl :: (GetKoln w b, PGOrd b) => (a -> w) -> Order a

Descending order, NULLs last.

WDef

data WDef a

Whether to write a DEFAUT value or a specific value into a database column.

WDef is isomorphic to Maybe. It exists mainly to avoid accidentally mixing the two of them together.

Constructors

WDef
Write DEFAULT.

WVal a
Write the specified value.

wdef :: b -> (a -> b) -> WDef a -> b

Case analysis for WDef.

Like maybe, evaluates to the first argument if WDef, otherwise applies the given function to the a in WVal.

Types

class (Tabla t1, Tabla t2, HasColName t1 c1, HasColName t2 c2, Database t1 ~ Database t2, Col_PgType (Col_ByName t1 c1) ~ Col_PgType (Col_ByName t2 c2)) => Comparable t1 c1 t2 c2

Provide Comparable instances for every two columns that you want to be able to compare (e.g., using eq).

class NotNullable x => PGType x

Only instances of PGType can be arguments to Kol or Koln

data Col name wd rn pgType hsType

Column description.

This is only used as a promoted datatype expected to have kind Col Symbol WD RN * *.

name: Column name.

wd: Whether DEFAULT can be written to this column (WD) or not (W).

rn: Whether NULL might be read from this column (RN) or not (R).

pgType: Type of the column value used in Opaleye queries (e.g., PGText, PGInt2).

hsType: Type of the column value used in Haskell outside Opaleye queries. Hint: don't use something like Maybe Bool here if you want to indicate that this is an optional Bool column. Instead, use Int here and RN in the rn field.

Notice that Col is very different from Kol and Koln: Kol and Koln are used at runtime for manipulating values stored in columns, Col is used to describe the properties of a column at compile time.

Constructors

Col name wd rn pgType hsType

data C c

Tag to be used alone or with Tagged for uniquely identifying a specific column in an unknown table.

Constructors

C

data Tabla t => T t

Tag to be used alone or with Tagged for uniquely identifying a specific table in a specific schema.

Constructors

T

data Tabla t => TC t c

Tag to be used alone or with Tagged for uniquely identifying a specific column in a specific table in a specific schema.

Constructors

TC

data RN

Whether to read a plain value or possibly NULL.

Constructors

R
Read plain value.

RN
Possibly read NULL.

data WD

Whether to write a specific value or possibly DEFAULT.

Constructors

W
Write a specific value.

WD
Possibly write DEFAULT. See WDef.

Individual columns

type Col_ByName t c = Col_ByName' c (Cols t)

Lookup a column in Tabla t by its name.

type family Col_Name col :: Symbol

Equations

Col_Name (Col n w r p h) = n

type family Col_PgRType col :: *

Equations

Col_PgRType (Col n w R p h) = Kol p

Col_PgRType (Col n w RN p h) = Koln p

type family Col_PgRNType col :: *

Equations

Col_PgRNType (Col n w r p h) = Koln p

type family Col_PgWType col :: *

Equations

Col_PgWType (Col n W r p h) = Col_PgRType (Col n W r p h)

Col_PgWType (Col n WD r p h) = WDef (Col_PgRType (Col n WD r p h))

type family Col_HsRType col :: *

Equations

Col_HsRType (Col n w R p h) = h

Col_HsRType (Col n w RN p h) = Maybe h

type family Col_HsIType col :: *

Equations

Col_HsIType (Col n W r p h) = Col_HsRType (Col n W r p h)

Col_HsIType (Col n WD r p h) = WDef (Col_HsRType (Col n WD r p h))

Set of columns

type Rec t xs = Tagged (T t) (Record xs)

All the representation of t used within opaleye-sot are (Rec t).

type Cols_HsR t = Map (Col_HsRFieldSym1 t) (Cols t)

Payload for (HsR t)

type Cols_HsI t = Map (Col_HsIFieldSym1 t) (Cols t)

Payload for (HsI t)

type Cols_PgR t = Map (Col_PgRSym1 t) (Cols t)

Payload for (PgR t)

type Cols_PgRN t = Map (Col_PgRNSym1 t) (Cols t)

Payload for (PgRN t)

type Cols_PgW t = Map (Col_PgWSym1 t) (Cols t)

Type of the Record columns when inserting or updating a row. Also, payload for (PgW t).

Miscellaneous

op1 :: Op1 a b fa fb xa xb => (fa -> fb) -> xa -> xb

Generalize the negative and positive arguments of the given function so that it works for as many combinations of (Kol x), (Koln x), (Tagged (TC t c) (Kol x)) or (Tagged (TC t c) (Koln x)) as possible.

op2 :: Op2 a b c fa fb fc xa xb xc => (fa -> fb -> fc) -> xa -> xb -> xc

Generalize the negative and positive arguments of the given function so that it works for as many combinations of (Kol x), (Koln x), (Tagged (TC t c) (Kol x)) or (Tagged (TC t c) (Koln x)) as possible. If xa and xb are Tagged with TC, then a Comparable constraint will be required on them.