Opaleye.SOT.Run

Documentation

commit :: (MonadIO m, MonadReader Connection m) => m ()

Query

runQueryMany

Arguments

:: (MonadIO m, MonadThrow m, MonadReader Connection m, Default QueryRunner v hs)

=> (hs -> Either SomeException r)

-> Query v

-> m [r]

Query and fetch zero or more resulting rows.

runQuery1

Arguments

:: (MonadIO m, MonadThrow m, MonadReader Connection m, Default QueryRunner v hs)

=> (hs -> Either SomeException r)

-> Query v

-> m (Maybe r)

Query and fetch zero or one resulting row.

Throws ErrTooManyRows if there is more than one row in the result.

runQueryHead

Arguments

:: (MonadIO m, MonadThrow m, MonadReader Connection m, Default QueryRunner v hs)

=> (hs -> Either SomeException r)

-> Query v

-> m r

Query and fetch one resulting row.

Throws ErrTooManyRows if there is more than one row in the result, and ErrNoRows if there is no row in the result.

Insert

runInsertMany

Arguments

:: (MonadIO m, MonadReader Connection m)

=> Table w v

-> [w]

-> m Int64

Insert zero or more rows.

runInsert1

Arguments

:: (MonadIO m, MonadReader Connection m)

=> Table w v

-> w

-> m Int64

Insert one row.

Returning

runInsertReturningMany

Arguments

:: (MonadIO m, MonadReader Connection m, Default QueryRunner v hs)

=> (hs -> Either SomeException r)

-> Table w v

-> w

-> m [r]

Insert zero or more rows, returning data from the rows actually inserted.

runInsertReturning1

Arguments

:: (MonadIO m, MonadThrow m, MonadReader Connection m, Default QueryRunner v hs)

=> (hs -> Either SomeException r)

-> Table w v

-> w

-> m (Maybe r)

Insert 1 row, returning data from the zero or one rows actually inserted.

Throws ErrTooManyRows if there is more than one row in the result.

runInsertReturningHead

Arguments

:: (MonadIO m, MonadThrow m, MonadReader Connection m, Default QueryRunner v hs)

=> (hs -> Either SomeException r)

-> Table w v

-> w

-> m r

Insert 1 row, returning data from the one row actually inserted.

Throws ErrTooManyRows if there is more than one row in the result, and ErrNoRows if there is no row in the result.

Update

runUpdate

Arguments

:: (MonadIO m, MonadReader Connection m, GetKol gkb PGBool)

=> Table w r

-> (r -> w)

-> (r -> gkb)

-> m Int64

Like Opaleye's runUpdate, but the predicate is expected to return a (GetKol w PGBool).

It is recommended that you use runUpdateTabla if you are trying to update a table that is an instance of Tabla. The result is the same, but the this function might be less convenient to use.

runUpdateTabla

Arguments

:: (Tabla t, MonadIO m, MonadReader Connection m, GetKol gkb PGBool)

=> T t

-> (PgW t -> PgW t)

-> (PgR t -> gkb)

-> m Int64

Like runUpdateTabla', but takes t explicitely for the times when it can't be inferred.

runUpdateTabla'

Arguments

:: (Tabla t, MonadIO m, MonadReader Connection m, GetKol gkb PGBool)

=> (PgW t -> PgW t)
Upgrade current values to new values.

-> (PgR t -> gkb)
Whether a row should be updated.

-> m Int64
Number of updated rows.

Like runUpdate, but specifically designed to work well with Tabla.

Exception

data ErrTooManyRows

Exception type indicating too many rows.

Constructors

ErrTooManyRows Int String
Number of rows, SQL string

data ErrNoRows

Exception type indicating no rows.

Constructors

ErrNoRows String
SQL string